requirements.txt
langgraph
langchain
langchain-xai
langchain-community
pygithub
streamlit
python-dotenv
docker
pydantic